Add "'cobalt_code': 1" to dom.gyp, fix Windows build
To make cobalt_code 1 work, the following changes are made:
- Added MSVC_PUSH_DISABLE_WARNING 4355 to anywhere "this" is used in base
  member initializer list.
- Disabled warning C4702: unreachable code in xtree.
- Added static_cast to all places where implicit conversion will cause warnings,
  or change variable type if it makes sense.
- Fixed all unreferenced parameter warnings.

b/22613037

+++ cobalt/dom/named_node_map.cc
@@ -31,7 +31,9 @@ NamedNodeMap::NamedNodeMap(const scoped_refptr<Element>& element)
   Stats::GetInstance()->Add(this);
 }
 
-unsigned int NamedNodeMap::length() const { return attribute_names_.size(); }
+unsigned int NamedNodeMap::length() const {
+  return static_cast<unsigned int>(attribute_names_.size());
+}
 
 scoped_refptr<Attr> NamedNodeMap::Item(unsigned int item) {
   if (item < attribute_names_.size()) {
